Function ic_canister_client::agent::sign_submit

pub fn sign_submit(
    content: HttpSubmitContent, 
    sender: &Sender
) -> Result<(HttpRequestEnvelope<HttpSubmitContent>, MessageId), Box<dyn Error>>

Wraps the content into an envelope that contains the message signature.

Prerequisite: content contains a sender field that is compatible with the keypair argument.