Struct ic_canister_client::agent::Agent

pub struct Agent {
    pub url: Url,
    pub sender: Sender,
    pub sender_field: Blob,
    // some fields omitted
}

An agent to talk to the Internet Computer through the public endpoints.

Fields
url: Url

Url of the replica to target. This should NOT contain a URL path like "/api/v1/submit".

sender: Sendersender_field: Blob

The values that any 'sender' field should have when issuing calls with the user corresponding to this Agent.

Implementations

impl Agent

pub fn new(url: Url, sender: Sender) -> Self

Creates an agent.

url: Url of the replica to target. This should NOT contain a URL path like "/api/v1/submit". It should contain a port, if needed.

The sender identifies the sender on whose behalf the requests are sent. If the requests are authenticated, the corresponding pub_key and sender_sig field are set in the request envelope.

pub fn new_with_client(
http_client: HttpClient,
url: Url,
sender: Sender
) -> Self

Creates an agent.

Same as above except gives the caller the option to retain a pre-existing client.

pub fn new_for_test(&self, sender: Sender) -> Self

This is needed by rust_canister tests

pub fn with_ingress_timeout(self, ingress_timeout: Duration) -> Self

Sets the timeout for ingress requests.

pub fn with_query_timeout(self, query_timeout: Duration) -> Self

Sets the timeout for queries.

pub fn with_install_timeout(self, install_timeout: Duration) -> Self

Sets the timeout for canister installation.

pub async fn query_cup_endpoint<'_>(
&'_ self,
param: Option<CatchUpPackageParam>
) -> Result<Option<CatchUpPackage>, String>

Queries the cup endpoint given the provided CatchUpPackageParams.

pub async fn execute_query<'_, '_, '_>(
&'_ self,
canister_id: &'_ CanisterId,
method: &'_ str,
arg: Vec<u8>
) -> Result<Option<Vec<u8>>, String>

Calls the query method 'method' on the given canister, optionally with 'arguments'.

pub async fn execute_update<'_, '_, S: ToString>(
&'_ self,
canister_id: &'_ CanisterId,
method: S,
arguments: Vec<u8>,
nonce: Vec<u8>
) -> Result<Option<Vec<u8>>, String>

Calls the query method 'method' on the canister located at 'url', optionally with 'arguments'.

pub async fn wait_ingress<'_, '_>(
&'_ self,
request_id: MessageId,
deadline: Instant,
canister_id: &'_ CanisterId
) -> Result<RequestStatus, String>

Requests the status of a pending canister update call request exactly once using the read_state API.

This is intended to be used in a loop until a final state is reached.

pub async fn ic_api_version<'_>(&'_ self) -> Result<String, String>

Requests the version of the public spec supported by this node by querying /api/v1/status.

pub async fn impl_version<'_>(&'_ self) -> Result<Option<String>, String>

Requests the Replica impl version of this node by querying /api/v1/status

pub async fn root_key<'_>(&'_ self) -> Result<Option<Blob>, String>

Requests the root key of this node by querying /api/v1/status

pub fn http_client(&self) -> &HttpClient

impl Agent

pub async fn install_canister<'_>(
&'_ self,
install_args: InstallCodeArgs
) -> Result<(), String>

impl Agent

pub fn prepare_update_raw<S: ToString>(
&self,
canister_id: &CanisterId,
method: S,
arguments: Vec<u8>,
nonce: Vec<u8>,
ingress_expiry: Time
) -> Result<(HttpRequestEnvelope<HttpSubmitContent>, MessageId), Box<dyn Error>>

Prepares an update request.

pub fn prepare_update<S: ToString>(
&self,
canister_id: &CanisterId,
method: S,
arguments: Vec<u8>,
nonce: Vec<u8>
) -> Result<(Vec<u8>, MessageId), Box<dyn Error>>

Prepares and serailizes a CBOR update request.

pub fn prepare_query(
&self,
canister_id: &CanisterId,
method: &str,
arguments: Vec<u8>
) -> Result<Vec<u8>, Box<dyn Error>>

Prepares and serializes a CBOR query request.

pub fn prepare_read_state(
&self,
paths: &[Path]
) -> Result<Vec<u8>, Box<dyn Error>>

Prepares and serializes a CBOR read_state request, with the given paths